refactor(services): relocate agentic workflow request formatter
