+++ libs/domains/services/feature/src/lib/service-creation-flow/agentic-workflow/agentic-workflow-request.ts
@@ -1,5 +1,5 @@
 import { type AgenticWorkflowHeader, type AgenticWorkflowRequest } from 'qovery-typescript-axios'
-import { type AgenticWorkflowFormData } from '../../service-creation-flow/agentic-workflow/agentic-workflow-context'
+import { type AgenticWorkflowFormData } from './agentic-workflow-context'
 
 function formatWhitelistHosts(value: string) {
   return value
